otel client via config
We expect you to provide an otel-collector output, and we will speak
http, https, grpc or insecure-grpc to that endpoint.
There's an easy path to use providing scrape-able Prometheus metrics
too, but that is not implemented!

We do not let you provide certificate auth, which might be a thing for
gRPC...

I'm awkwardly reusing the `[statsd]` section of the config file - I
didn't want to introduce a sibling `[stats]` section, but maybe this
should all go into a `[otel]` section to avoid confusion? :shrug: